Fix money round getting info from DB Fix performance on search with big list of movements

+++ data/src/main/java/com/android/data/local/model/BalanceVO.kt
@@ -2,23 +2,22 @@ package com.android.data.local.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import java.math.BigDecimal
 
 @Entity(tableName = "Balance")
 data class BalanceVO(
     @PrimaryKey
     val id: Int = 0,
-    val IngresosEfectivo: BigDecimal,
-    val EgresosEfectivo: BigDecimal,
-    val IngresosElectronico: BigDecimal,
-    val EgresosElectronico: BigDecimal,
-    val Retiros: BigDecimal,
-    val ComprasTC: BigDecimal,
-    val TengoEfectivo: BigDecimal,
-    val TengoElectronico: BigDecimal,
-    val TengoTotal: BigDecimal,
-    val DiferenciaIngresos: BigDecimal,
-    val DiferenciaEgresos: BigDecimal,
-    val TotalIngresos: BigDecimal,
-    val TotalEgresos: BigDecimal
+    val IngresosEfectivo: Double,
+    val EgresosEfectivo: Double,
+    val IngresosElectronico: Double,
+    val EgresosElectronico: Double,
+    val Retiros: Double,
+    val ComprasTC: Double,
+    val TengoEfectivo: Double,
+    val TengoElectronico: Double,
+    val TengoTotal: Double,
+    val DiferenciaIngresos: Double,
+    val DiferenciaEgresos: Double,
+    val TotalIngresos: Double,
+    val TotalEgresos: Double
 )
